Added Base refill rate limiter
